Created New DashboadTimer
created new dashboard timer and testing inventory adjsutment
