+++ nbproject/private/private.xml
@@ -23,7 +23,9 @@
         <group name="GCard"/>
         <group name="POS"/>
         <group>
-            <file>file:/D:/Jovan/GitHub/cas-food-inventory-fx/src/org/rmj/cas/food/inventory/fx/views/POReceivingController.java</file>
+            <file>file:/D:/Jovan/GitHub/cas-food-inventory-fx/src/org/rmj/cas/food/inventory/fx/views/InvTransferController.java</file>
+            <file>file:/D:/Jovan/GitHub/cas-food-inventory-fx/src/org/rmj/cas/food/inventory/fx/views/TimerDashBoard.java</file>
+            <file>file:/D:/Jovan/GitHub/cas-food-inventory-fx/src/org/rmj/cas/food/inventory/fx/views/InvAdjustmentController.java</file>
         </group>
     </open-files>
 </project-private>
